update to new fmt/format version

+++ src/helics/apps/Recorder.cpp
@@ -10,8 +10,7 @@ SPDX-License-Identifier: BSD-3-Clause
 #include "../application_api/Filters.hpp"
 #include "../application_api/queryFunctions.hpp"
 #include "../common/JsonProcessingFunctions.hpp"
-#include "../common/fmt_format.h"
-#include "../common/fmt_ostream.h"
+
 #include "../core/helicsCLI11.hpp"
 #include "PrecHelper.hpp"
 #include "gmlc/utilities/base64.h"
@@ -20,6 +19,8 @@ SPDX-License-Identifier: BSD-3-Clause
 #include <algorithm>
 #include <fstream>
 #include <iostream>
+#include <fmt/format.h>
+#include <fmt/ostream.h>
 #include <map>
 #include <regex>
 #include <set>
@@ -389,20 +390,20 @@ void Recorder::captureForCurrentTime(Time currentTime, int iteration)
                 if (val.size() < 150) {
                     if (iteration > 0) {
                         valstr = fmt::format(
-                            "[{}:{}]value {}={}", currentTime, iteration, sub.getTarget(), val);
+                            "[{}:{}]value {}={}", static_cast<double>(currentTime), iteration, sub.getTarget(), val);
                     } else {
-                        valstr = fmt::format("[{}]value {}={}", currentTime, sub.getTarget(), val);
+                        valstr = fmt::format("[{}]value {}={}", static_cast<double>(currentTime), sub.getTarget(), val);
                     }
                 } else {
                     if (iteration > 0) {
                         valstr = fmt::format("[{}:{}]value {}=block[{}]",
-                                             currentTime,
+                            static_cast<double>(currentTime),
                                              iteration,
                                              sub.getTarget(),
                                              val.size());
                     } else {
                         valstr = fmt::format("[{}]value {}=block[{}]",
-                                             currentTime,
+                            static_cast<double>(currentTime),
                                              sub.getTarget(),
                                              val.size());
                     }
@@ -425,13 +426,13 @@ void Recorder::captureForCurrentTime(Time currentTime, int iteration)
                 std::string messstr;
                 if (mess->data.size() < 50) {
                     messstr = fmt::format("[{}]message from {} to {}::{}",
-                                          currentTime,
+                        static_cast<double>(currentTime),
                                           mess->source,
                                           mess->dest,
                                           mess->data.to_string());
                 } else {
                     messstr = fmt::format("[{}]message from {} to {}:: size {}",
-                                          currentTime,
+                        static_cast<double>(currentTime),
                                           mess->source,
                                           mess->dest,
                                           mess->data.size());
